+++ Models/Storno.cs
@@ -25,6 +25,6 @@ namespace CarLeasingViewer.Models
 
         public int DayIndexEnd { get; private set; }
 
-        public double DayCount { get; set; }
+        public decimal DayCount { get; set; }
     }
 }
